+++ tools/xdhwebq/htdocs/prolog.php
@@ -26,9 +26,8 @@ $out = str_replace(array("\r", "\n", "\t"), '','
 			<!-- BEGIN -->
 			<meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
 			<meta http-equiv="Cache-Control" content="no-cache"/>
-			<!--script src="faas/xdhtml_20200804.js"></script-->
-			<script src="faas/xdhtml_20200821.js"></script>
-			<script src="faas/xdhwebq_20200608.js"></script>
+			<script src="faas/xdhtml.js"></script>
+			<script src="faas/xdhwebq.js"></script>
 			<script src="faas/qrcode.min.js"></script>
 			<style id="XDHStyle">
 				.xdh_style {